% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_term.R
\name{add_term}
\alias{add_term}
\title{add_term}
\usage{
add_term(x, term)
}
\arguments{
\item{x}{object to have term appended}

\item{term}{unquoted term}
}
\description{
Add term to an existing object
}
\details{
Adds a term to a an object.  

When `x` is a formula, the term is added to the RHS of the 
formula.
}
\examples{

x <- call( '+', as.name('x1'), as.name('x2'))

add_term(x, x3)
add_term(x, -x3)

add_term(x, 'x3')
# add_term(x, "-x3")
add_term(x, x3+x4 )
add_term( y ~ x1 + x2, x3 )

}
