% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env}
\alias{env}
\alias{env.formula}
\title{Get the environment}
\usage{
env(x)

\method{env}{formula}(x)
}
\arguments{
\item{x}{object to get environoment from}
}
\value{
Environment
}
\description{
Get the environment
}
\details{
S3 returns the environment associated with an object.

For a formula object, `env` returns the environment in the `.Environment` 
attribute.
}
\examples{

env( lhs ~ rhs ) 

}
