% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{lhs}
\alias{.lhs.singular}
\alias{.replace.lhs.plural}
\alias{.replace.lhs.singular}
\alias{.replace.op.plural}
\alias{.replace.rhs.plural}
\alias{.replace.rhs.singular}
\alias{.replace.ths.plural}
\alias{.rhs.singular}
\alias{lhs}
\alias{lhs,<--method}
\alias{lhs,call-method}
\alias{lhs,formula-method}
\alias{lhs,list-method}
\alias{lhs<-}
\alias{lhs<-,call-method}
\alias{lhs<-,expression-method}
\alias{lhs<-,formula-method}
\alias{lhs<-,list-method}
\alias{op}
\alias{op,<--method}
\alias{op,call-method}
\alias{op,expression-method}
\alias{op,formula-method}
\alias{op,list-method}
\alias{op,name-method}
\alias{op<-}
\alias{op<-,call-method}
\alias{op<-,expression-method}
\alias{op<-,formula-method}
\alias{op<-,list-method}
\alias{rhs}
\alias{rhs,<--method}
\alias{rhs,call-method}
\alias{rhs,call-methods}
\alias{rhs,formula-method}
\alias{rhs,list-method}
\alias{rhs<-}
\alias{rhs<-,call-method}
\alias{rhs<-,expression-method}
\alias{rhs<-,formula-method}
\alias{rhs<-,list-method}
\title{Get/set left- or right- hand side of a formula, call, expression, etc.}
\usage{
lhs(x, ...)

.lhs.singular(x)

\S4method{lhs}{call}(x)

\S4method{lhs}{formula}(x)

\S4method{lhs}{`<-`}(x)

\S4method{lhs}{list}(x, ...)

lhs(x) <- value

.replace.lhs.singular(x, value)

.replace.lhs.plural(x, value)

op(x)

\S4method{op}{formula}(x)

\S4method{op}{call}(x)

\S4method{op}{name}(x)

\S4method{op}{expression}(x)

\S4method{op}{list}(x)

\S4method{op}{`<-`}(x)

op(x) <- value

.replace.op.plural(x, value)

rhs(x, ...)

.rhs.singular(x)

\S4method{rhs}{call}(x)

\S4method{rhs}{formula}(x)

\S4method{rhs}{`<-`}(x)

\S4method{rhs}{list}(x, ...)

rhs(x) <- value

.replace.rhs.singular(x, value)

.replace.rhs.plural(x, value)
}
\value{
the left- or right- hand side of \code{x}
}
\description{
Get/set left- or right- hand side of a formula, call, expression, etc.
}
\examples{
#' -tk
}

