\name{.invert.single}
\alias{.invert.plural}
\alias{.invert.single}
\title{Invert multiple elements of a multiple element object}
\usage{
.invert.single(x)

.invert.plural(x)
}
\arguments{
  \item{x}{object to invert from}
}
\description{
Invert multiple elements of a multiple element object
}
\seealso{
.invert.single
}

