\name{operators}
\Rdversion{1.1}
\alias{relational.operators}
\alias{relational.inverses}
\alias{tilde}
\alias{tilde.operator}

\docType{data}
\title{
  Internally used data.
}

\description{
  Documentation of internal data structures:
 
}

\format{

  relational.operators vector of relational operators: 
    c( '<', '>', '<=', '>=', '==', '!=' )

  tilde,tilde-operator '~'.  It may include other operators laters 

  op.types is a hash with operators as keys and the type of opertors as values

  inverse.operators is a hash with operators as keys and the inverse of that
  operator as values.  

}

\details{
}


\keyword{datasets}

