% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.terms.R
\name{split.terms}
\alias{split.terms}
\title{Split object into terms}
\usage{
split.terms(x, recursive = FALSE)
}
\arguments{
\item{x}{object to split terms from}

\item{recursive}{logical; whether to split terms recursively in parenthetical
enclosed terms  (depth-first).  (Default: FALSE)}
}
\value{
experession vector of terms
}
\description{
Split formulas, call and expressions into terms.
}
\details{
Splits into 
See **examples**.
}
\examples{
  
  split.terms(1)                   # 1
  split.terms( quote(a) )          # a     
  split.terms( quote(a+1) )        # a, 1
  split.terms( quote(1+a) )        # 1, a
  split.terms( quote(a+b+c) )      # a, b, c
  split.terms( quote((a+b)+1) )    # (a+b),1
  split.terms( quote((a+b)+1), recursive=TRUE )    # (a+b),1
  
  split.terms( quote(a-1) )
  split.terms( quote(-a-1))
}
\references{
* [SO: How to split a formula](https://stackoverflow.com/questions/39155701/how-to-split-a-formula-in-r)
}
