% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/terms.R
\name{terms.call}
\alias{terms}
\alias{terms.call}
\alias{terms.expression}
\title{terms}
\usage{
\method{terms}{call}(x, ...)

\method{terms}{expression}(x, ...)
}
\arguments{
\item{x}{A call object}

\item{...}{Arguments passed to \code{\link{terms.formula}}

This S3 method returns a terms object for a call methods using a dispatch to
\code{ terms.formula }.

The terms are generated by making a rhs only call to \code{ terms.formula }.

\code{data} is only needed and must be explicitly specified, i.e.  \code{
data = } if there are special elements such as '.'.  Otherwise the
\code{data} argument is unused.

Some edge cases may not be supported.}
}
\value{
A terms object.  See \code{\link{terms.object}} for details.
}
\description{
terms method for call and expression objects
}
\examples{
terms( quote( A + B ) )

    data(iris)
    x <- terms( quote( . - Species ) , data=iris )
}
\author{
Christopher Brown
}
\seealso{
\code{\link{terms.object}} and \code{\link{terms.formula}}
}
\keyword{manip}
\keyword{symbolmath}
\keyword{utilities}

