% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_sign.R
\name{toggle_sign}
\alias{toggle_sign}
\title{toggle the sign of an expression}
\usage{
toggle_sign(x)
}
\arguments{
\item{x}{expression}
}
\description{
toggle the sign of an expression
}
\details{
`toggle_sign` changes the sign of an expression for `+` tp `-` and 
visa-versa.
}
\examples{

  toggle_sign(1:3)
  toggle_sign( quote(a) )
  toggle_sign( quote(-a) )
  
  exp <- expression( a, -b, -(a-b) )
  toggle_sign(exp)
  
}
