% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.character.formula.R
\name{as.character.formula}
\alias{as.character.formula}
\alias{as.character}
\title{Converts a formula to character}
\usage{
\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{formula object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A character vector
}
\description{
Convers a formula to character representaion
}
\details{
Coerces formula to a character by deparsing.
}
\examples{

  as.character( y ~ mx +  b )

## The function is currently defined as
function(x)
  Reduce( paste, deparse(x) )
  
}
\seealso{
\code{\link[base]{deparse}}
}
\author{
Christopher Brown
}
\keyword{manip}
\keyword{utilities}
