\name{formula.tools-package}
\Rdversion{1.1}
\alias{formula.tools-package}
\alias{formula.tools}
\docType{package}
\title{
   Tools for working with formulas, expressions, calls and other R objects
}

\description{
  This packages provides a number of useful utilities for working with formulas, expressions, calls, names, symbols and other R objects.  This utilities can be used for symbolic manupulation of these objects. 
}
\details{
\tabular{ll}{
    Package: \tab formtools\cr
    Type: \tab Package\cr
    Version: \tab 0.14.1\cr
    Date: \tab 2009-11-28\cr
    License: \tab LPGL(>=2)\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
Christopher Brown  
Maintainer: Christopher Brown <cbrown@opendatagroup.com>
\url{http://www.opendatagroup.com}
}

\keyword{ package }

\seealso{
}

\examples{
}
