% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/invert.R
\name{.invert.single}
\alias{.invert.plural}
\alias{.invert.single}
\title{Invert multiple elements of a multiple element object}
\usage{
.invert.single(x)

.invert.plural(x)
}
\arguments{
\item{x}{object to invert from}
}
\description{
Invert multiple elements of a multiple element object
}
\seealso{
.invert.single
}

