% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.type.R
\docType{methods}
\name{op.type}
\alias{op.type}
\alias{op.type-methods}
\alias{op.type,call-method}
\alias{op.type,formula-method}
\alias{op.type,<--method}
\alias{op.type,ANY-method}
\alias{op.type,expression-method}
\alias{op.type,list-method}
\title{Get the operator type used in an call, formula, expression, etc.}
\usage{
op.type(x)

\S4method{op.type}{call}(x)

\S4method{op.type}{formula}(x)



\S4method{op.type}{ANY}(x)

\S4method{op.type}{expression}(x)

\S4method{op.type}{list}(x)
}
\arguments{
\item{x}{object from which to extract the operator type}
}
\value{
a character vector of the operator type(s)
}
\description{
Get the operator type used in an call, formula, expression, etc.
}
\seealso{
\code{\link{op}}, \code{\link[operator.tools]{operator.type}}
}
